+++ GameServer/GameServer/world/mapelement.h
@@ -7,11 +7,11 @@ class Rectangle : public Serializable {
 public:
 	int width;
 	int height;
-	int x;
-	int y;
-	int rotation;
+	float x;
+	float y;
+	float rotation;
 
-	Rectangle(int width, int height, int x, int y, int rotation = 0);
+	Rectangle(int width, int height, float x, float y, float rotation = 0);
 	virtual std::string Serialize();
 
 protected:
@@ -22,7 +22,7 @@ class MapElement : public Rectangle {
 public:
 	int layer;
 
-	MapElement(int width, int height, int x, int y, int rotation = 0, int layer = 0);
+	MapElement(int width, int height, float x, float y, int rotation = 0, int layer = 0);
 	virtual ~MapElement() {}
 	virtual std::string GetType() = 0;
